Upgrade laravel and laravel-mix

+++ database/migrations/2017_04_14_013622_create_votes_table.php
@@ -13,8 +13,8 @@ class CreateVotesTable extends Migration
     public function up()
     {
         Schema::create('votes', function (Blueprint $table) {
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('votable_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('votable_id');
             $table->string('votable_type')->index();
             $table->enum('type', ['up_vote', 'down_vote'])->default('up_vote');
